feat: team and advisors changed to toggle bootstrap on mobile devices
A variable isMobile returns either true or false. If it is not mobile, the original markup is generated with the content
If variable is true, the markup utilizes bootstrap's toggle features instead of hover tooltips.
The new markup does not use bootstraps md-5 column feature anymore, since toggling with this setup would cause float breakpoints to jump, thus the layout looked like a lego stuck to another lego but only on one side
To accomodate, the parent div has the class of 'flex-container' which applies flexbox stylings to imitate the original layout of the children.